feat: convert to typescript
Converts this module to typescript and uses aegir to build and run
tests.
